nit: turn int osize, nsize into size_t

+++ VM/include/lua.h
@@ -454,7 +454,7 @@ struct lua_Callbacks
     void (*debuginterrupt)(lua_State* L, lua_Debug* ar); // gets called when thread execution is interrupted by break in another thread
     void (*debugprotectederror)(lua_State* L);           // gets called when protected call results in an error
 
-    void (*onallocate)(lua_State* L, int osize, int nsize); // gets called when memory is allocated
+    void (*onallocate)(lua_State* L, size_t osize, size_t nsize); // gets called when memory is allocated
 };
 typedef struct lua_Callbacks lua_Callbacks;
